trim down the PR
